Fix mouse input type mismatch and FP optimizer bug
The mouse input was broken due to two issues:

1. Type mismatch between definition and declarations:
   - sys_amiga.c defined mouseX/mouseY as 'int' (32-bit)
   - in_amiga.c and vid_amiga.c declared them as 'short int' (16-bit)
   - This caused code to read wrong bytes from memory

2. GCC FP optimizer bug at -O2:
   - Mouse sensitivity calculations use FP multiplication
   - Similar to mathlib.c angle calculation issues
   - Compiling in_amiga.c with -O1 prevents misoptimization

Changes:
- Fixed type declarations: extern int mouseX, mouseY (was short int)
- Added -O1 compilation for in_amiga.c in Makefile
- Updated README with mouse movement bug documentation
- Updated optimization notes to mention both files needing -O1

+++ src/in_amiga.c
@@ -20,9 +20,9 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "quakedef.h"
 
-// Mouse globals set by vid_amiga.c
-extern short int mouseX;
-extern short int mouseY;
+// Mouse globals set by sys_amiga.c
+extern int mouseX;
+extern int mouseY;
 extern qboolean mouse_has_moved;
 
 cvar_t	m_filter = {"m_filter","1"};
